pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "numpy>=2.3.1",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
    "sentence-transformers>=5.0.0",
    "tweepy>=4.16.0",
]
